apps/api/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["SM Tanimur Rahman <smtanimurrahman404@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
sqlmodel = "^0.0.21"
alembic = "^1.13.2"
pydantic = "^2.8.2"
sqlalchemy = "^2.0.32"
asyncpg = "^0.29.0"
psycopg2 = "^2.9.9"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"
bcrypt = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
start= "app.main:start"
